3131 Дать пользователю возможность редактировать собственные данные http://task.artektiv.ru/dev/view/id/3131

+++ protected/models/User.php
@@ -24,6 +24,7 @@ class User extends CActiveRecord
 {
 	public $active = 1;
 	public $repassword;
+	public $old_password;
 
 	public $rememberMe = false;
 	private $_identity = null;
@@ -66,13 +67,15 @@ class User extends CActiveRecord
 		return array(
 			array('email, password', 'required', 'on' => 'login'),
 			array('email, name, surname, companies_ids', 'required', 'on' => 'insert,update,owner_update,owner_create'),
+			array('email, name, surname', 'required', 'on' => 'profile'),
 			array('companies_ids', 'safe', 'on' => 'owner_update,only_company,owner_create'),
-			array('email', 'unique', 'on' => 'insert,update,owner_update,owner_create'),
-			array('email', 'email', 'on' => 'insert,update,owner_update,owner_create'),
+			array('email', 'unique', 'on' => 'insert,update,owner_update,owner_create,profile'),
+			array('email', 'email', 'on' => 'insert,update,owner_update,owner_create,profile'),
 			array('active', 'numerical', 'integerOnly'=>true, 'on' => 'insert,update,owner_update,owner_create'),
-			array('password, repassword', 'required', 'on' => 'insert, ch_pass,owner_create'),
-			array('repassword', 'compare', 'compareAttribute' => 'password', 'on' => 'insert, ch_pass, owner_create'),
-			array('phone', 'safe', 'on' => 'insert,update,owner_update,owner_create'),
+			array('old_password', 'required', 'on' => 'change_pass'),
+			array('password, repassword', 'required', 'on' => 'insert, change_pass,owner_create'),
+			array('repassword', 'compare', 'compareAttribute' => 'password', 'on' => 'insert, change_pass, owner_create'),
+			array('phone', 'safe', 'on' => 'insert,update,owner_update,owner_create,profile'),
 			// The following rule is used by search().
 			// Please remove those attributes that should not be searched.
 			array('id, email, password, name, surname, phone, active', 'safe', 'on'=>'search'),
@@ -100,6 +103,7 @@ class User extends CActiveRecord
 		return array(
 			'id' => 'ID',
 			'email' => 'E-mail',
+			'old_password' => 'Старый пароль',
 			'password' => 'Пароль',
 			'repassword' => 'Повтор пороля',
 			'name' => 'Имя',
@@ -135,27 +139,29 @@ class User extends CActiveRecord
 
 	protected function afterSave()
 	{
-		$companies_ids = array();
-		if (in_array($this->scenario, array('insert', 'update'))) {
-			User2company::model()->deleteAllByAttributes(array('user_id' => $this->id));
-		} elseif (in_array($this->scenario, array('owner_update', 'only_company', 'owner_create'))) {
-			foreach ($this->user2company as $u2c) {
-				$companies_ids[] = $u2c->company_id;
-				//TODO костыль :(
-				if ($u2c->company->admin_user_id == Yii::app()->user->id) {
-					$u2c->delete();
+		if (!in_array($this->scenario, array('profile', 'change_pass'))) {
+			$companies_ids = array();
+			if (in_array($this->scenario, array('insert', 'update'))) {
+				User2company::model()->deleteAllByAttributes(array('user_id' => $this->id));
+			} elseif (in_array($this->scenario, array('owner_update', 'only_company', 'owner_create'))) {
+				foreach ($this->user2company as $u2c) {
+					$companies_ids[] = $u2c->company_id;
+					//TODO костыль :(
+					if ($u2c->company->admin_user_id == Yii::app()->user->id) {
+						$u2c->delete();
+					}
 				}
 			}
-		}
 
-		foreach ($this->companies_ids as $c_id) {
-			$uc = new User2company();
-			$uc->user_id = $this->id;
-			$uc->company_id = $c_id;
-			if ($this->isNewRecord || in_array($this->scenario, array('insert', 'update'))) {
-				$uc->save();
-			} elseif ($this->create_user_id == Yii::app()->user->id or in_array($c_id, $companies_ids)) { //TODO еще один
-				$uc->save();
+			foreach ($this->companies_ids as $c_id) {
+				$uc = new User2company();
+				$uc->user_id = $this->id;
+				$uc->company_id = $c_id;
+				if ($this->isNewRecord || in_array($this->scenario, array('insert', 'update'))) {
+					$uc->save();
+				} elseif ($this->create_user_id == Yii::app()->user->id or in_array($c_id, $companies_ids)) { //TODO еще один
+					$uc->save();
+				}
 			}
 		}
 		parent::afterSave();
